requirements.txt
mysql-connector
uvicorn[standard]
fastapi
python-multipart
fastapi-sessions
bcrypt
jinja2